backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Bcrypt
Flask-Migrate
python-dotenv
psycopg2-binary
langchain
langchain-groq
requests
